The basic framework in all its tarnished glory.
Now to make it do things! Obviously the Graph is central, but I'd like to keep
it generic enough to export later, as a kind of light-weight RGL. I think it's
pretty neat that a Runner just pipelines the loader, analyser and formatter,
but of course, this could be improved, and it seems somewhat redundant to have
a single-method class. The may "RailsX" modules are probably overkill, too.
Mweh, let's see.